requirements.txt
asgiref==3.8.1
Django~=5.2
djangorestframework==3.15.1
psycopg==3.1.18
python-decouple==3.8
python-dotenv==1.2.1